Use a single exit point
This turns ConverterError into a real exception and stops exiting the
program from other places like main() which is a requirement for usage
of these classes in other applications (you don't want your whole app to
exit abort when importing iptables fails).

This will allow us to turn iptables_converter into a module in the
future.